microskel/requirements.txt
injector
Flask
Flask-SQLAlchemy
sqlalchemy
flask-injector
python-consul
python-decouple
mysqlclient
python-logstash
flask-cors
requests